Toast fixes and improvements
